requirement.txt
natsort
intervaltree
argparse
path
pandas
numpy
biopython
gffutils
gzip
rich